requirements.txt
dnspython==2.2.1
fastapi==0.78.0
httptools==0.5.0
imblearn==0.0
pip-chill==1.0.1
#python-dotenv==0.21.0
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
wincertstore==0.2
xgboost==1.6.2
pandas
PyYAML
numpy
pandas
scikit-learn
apache-airflow
-e .